re-organization, new ObjectID format with hacked compound domain/email index

+++ credshed/lib/db.py
@@ -6,6 +6,7 @@ import copy
 import time
 import queue
 import pymongo
+import configparser
 from .leak import *
 from .util import *
 from .errors import *
@@ -19,31 +20,27 @@ from subprocess import run, PIPE
 
 class DB():
 
-    def __init__(self):
+    def __init__(self, use_metadata=True):
+
+        self.config = self.parse_config()
 
         try:
 
-            ### MONGO ###
+            ### MONGO PRIMARY ###
+
+            main_server = self.config['MONGO PRIMARY']['server']
+            main_port = int(self.config['MONGO PRIMARY']['port'])
+            main_db = self.config['MONGO PRIMARY']['db']
 
             # main DB
-            self.main_client = pymongo.MongoClient('127.0.0.1', 27000)
-            self.main_db = self.main_client['credshed']
+            self.main_client = pymongo.MongoClient(main_server, main_port)
+            self.main_db = self.main_client[main_db]
 
-            # accounts
-            try:
-                self.main_db.create_collection('accounts')
-            except pymongo.errors.CollectionInvalid:
-                pass
+            #try:
+            #    self.main_db.create_collection('accounts')
+            #except pymongo.errors.CollectionInvalid:
+            #    pass
             self.accounts = self.main_db.accounts
-
-            # meta DB (account metadata including source information, counters, leak <--> account associations, etc.)
-            self.meta_client = pymongo.MongoClient('127.0.0.1', 27001)
-            self.meta_db = self.meta_client['credshed']
-
-            try:
-                self.meta_db.create_collection('account_tags')
-            except pymongo.errors.CollectionInvalid:
-                pass
 
         except pymongo.errors.PyMongoError as e:
             error = str(e) + '\n'
@@ -53,18 +50,44 @@ class DB():
                 pass
             raise CredShedDatabaseError(error)
 
+        self.use_metadata = False
+        if use_metadata:
+            try:
 
-        self.account_tags = self.meta_db.account_tags
+                ### MONGO METADATA ###
 
-        #self.accounts.create_index([('id', pymongo.ASCENDING)])
+                meta_server = self.config['MONGO METADATA']['server']
+                meta_port = int(self.config['MONGO METADATA']['port'])
+                meta_db = self.config['MONGO METADATA']['db']
+
+                # meta DB (account metadata including source information, counters, leak <--> account associations, etc.)
+                self.meta_client = pymongo.MongoClient(meta_server, meta_port)
+                self.meta_db = self.meta_client[meta_db]
+
+                #try:
+                #    self.meta_db.create_collection('account_tags')
+                #except pymongo.errors.CollectionInvalid:
+                #    pass
+                self.account_tags = self.meta_db.account_tags
+                self.use_metadata = True
+
+            except pymongo.errors.PyMongoError as e:
+                error = str(e) + '\n'
+                try:
+                    error += str(e.details)
+                except AttributeError:
+                    pass
+
+                errprint('[!] Problem with metadata database at {}:{}\n{}'.format(\
+                    meta_server, meta_port, error))
+
         #self.accounts.create_index([('username', pymongo.ASCENDING)], sparse=True, background=True)
-        #self.accounts.create_index([('domain', pymongo.ASCENDING)], sparse=True, background=True)
-        #self.accounts.create_index([('password', pymongo.ASCENDING), ('username', pymongo.ASCENDING)], sparse=True, background=True)
+        #self.accounts.create_index([('email', pymongo.ASCENDING)], sparse=True, background=True)
 
         # sources
-        self.sources = self.meta_db.sources
+        self.sources = self.main_db.sources
         # counters
-        self.counters = self.meta_db.counters
+        self.counters = self.main_db.counters
 
         # leak-specific counters
         self.leak_unique = 0
@@ -110,18 +133,21 @@ class DB():
 
             if query_type == 'email':
                 try:
+                    # _id begins with reversed domain
                     email, domain = keyword.lower().split('@')[:2]
-                    email = re.escape(email)
-                    domain = re.escape(domain[::-1])
-                    query_regex = r'^{}.*'.format(domain)
-                    query = {'$and': [{'email': email}, {'_id': {'$regex': query_regex}}]}
+                    domain_chunk = re.escape(domain[::-1])
+                    email_hash = decode(base64.b64encode(hashlib.sha256(encode(email)).digest()[:6]))
+                    query_str = domain_chunk + '|' +  email_hash
+
+                    query_regex = r'^{}.*'.format(query_str)
+                    query = {'_id': {'$regex': query_regex}}
                     errprint(query)
                     results['emails'] = self.accounts.find(query).limit(max_results)
                     #results['emails'] = self.accounts.find({'email': email, '_id': {'$regex': domain_regex}})
                     #results['emails'] = self.accounts.find({'email': email})
 
                 except ValueError:
-                    # assume email only
+                    # assume email without domain
                     email = r'^{}$'.format(keyword.lower())
                     query = {'email': {'$regex': email}}
                     errprint(query)
@@ -356,44 +382,48 @@ class DB():
 
     def remove_leak(self, source_id, batch_size=10000):
 
-        source = self.get_source(source_id)
-        accounts_deleted = 0
-        to_delete = []
+        if not self.use_metadata:
+            raise CredShedMetadataError('Removing leaks requires access to metadata. No metadata database is currently attached.')
 
-        errprint('\n[*] Deleting leak "{}{}"'.format(source.name, ':{}'.format(source.hashtype) if source.hashtype else ''))
+        else:
+            source = self.get_source(source_id)
+            accounts_deleted = 0
+            to_delete = []
 
-        try:
+            errprint('\n[*] Deleting leak "{}{}"'.format(source.name, ':{}'.format(source.hashtype) if source.hashtype else ''))
 
-            # delete accounts
-            for result in self.account_tags.find({'s': [source_id]}, {'_id': 1}):
-                to_delete.append(pymongo.DeleteOne(result))
-                if len(to_delete) % batch_size == 0:
+            try:
+
+                # delete accounts
+                for result in self.account_tags.find({'s': [source_id]}, {'_id': 1}):
+                    to_delete.append(pymongo.DeleteOne(result))
+                    if len(to_delete) % batch_size == 0:
+                        accounts_deleted += self.accounts.bulk_write(to_delete, ordered=False).deleted_count
+                        to_delete.clear()
+                        errprint('\r[+] Deleted {:,} accounts'.format(accounts_deleted), end='')
+
+                if to_delete:
                     accounts_deleted += self.accounts.bulk_write(to_delete, ordered=False).deleted_count
-                    to_delete.clear()
-                    errprint('\r[+] Deleted {:,} accounts'.format(accounts_deleted), end='')
 
-            if to_delete:
-                accounts_deleted += self.accounts.bulk_write(to_delete, ordered=False).deleted_count
+                # delete out of tags collection
+                self.account_tags.delete_many({'s': [source_id]})
+                # pull source ID from affected accounts
+                self.account_tags.update_many({'s': source_id}, {'$pull': {'s': source_id}})
 
-            # delete out of tags collection
-            self.account_tags.delete_many({'s': [source_id]})
-            # pull source ID from affected accounts
-            self.account_tags.update_many({'s': source_id}, {'$pull': {'s': source_id}})
+                errprint('\r[+] Deleted {:,} accounts'.format(accounts_deleted), end='')
 
-            errprint('\r[+] Deleted {:,} accounts'.format(accounts_deleted), end='')
-
-            self.sources.delete_many({'_id': source_id})
-            self.counters.update_one({'collection': 'sources'}, {'$unset': {str(source_id): ''}})
+                self.sources.delete_many({'_id': source_id})
+                self.counters.update_one({'collection': 'sources'}, {'$unset': {str(source_id): ''}})
 
 
-        except TypeError as e:
-            errprint(str(e))
-            errprint('[!] Can\'t find source "{}:{}"'.format(source.name, source.hashtype))
+            except TypeError as e:
+                errprint(str(e))
+                errprint('[!] Can\'t find source "{}:{}"'.format(source.name, source.hashtype))
 
-        errprint('\n[*] {:,} accounts deleted'.format(accounts_deleted))
-        errprint('[*] Done')
-        # sleep(1)
-        return accounts_deleted
+            errprint('\n[*] {:,} accounts deleted'.format(accounts_deleted))
+            errprint('[*] Done')
+            # sleep(1)
+            return accounts_deleted
 
 
     def add_source(self, source):
@@ -475,13 +505,14 @@ class DB():
                     stats.append('')
 
         except pymongo.errors.OperationFailure:
-            stats.append('[!] No accounts added yet', end='\n\n')
+            stats.append('[!] No accounts added yet\n')
 
         if db:
             db_stats = self.main_db.command('dbstats', scale=1048576)
             stats.append('[+] DB Stats (MB):')
             for k in db_stats:
-                stats.append('\t{}: {}'.format(k, db_stats[k]))
+                if k not in ['raw']:
+                    stats.append('\t{}: {}'.format(k, db_stats[k]))
 
         return '\n'.join(stats)
 
@@ -505,7 +536,9 @@ class DB():
     def account_count(self):
 
         try:
-            num_accounts_in_db = self.main_db.command('collstats', 'accounts', scale=1048576)['count']
+            collstats = self.main_db.command('collstats', 'accounts', scale=1048576)
+            print(collstats)
+            num_accounts_in_db = collstats['count']
         except KeyError:
             num_accounts_in_db = 0
 
@@ -521,10 +554,27 @@ class DB():
             return None
 
 
+    def parse_config(self):
+
+        # parse config file
+        config_filename = Path(__file__).resolve().parent.parent / 'credshed.config'
+        if not config_filename.is_file():
+            raise CredShedConfigError('Unable to find credshed config at {}'.format(config_filename))
+
+        config = configparser.ConfigParser()
+        config.read(str(config_filename))
+
+        return config
+
+
+
     def close(self):
 
-        self.main_client.close()
-        self.meta_client.close()
+        try:
+            self.main_client.close()
+            self.meta_client.close()
+        except AttributeError:
+            pass
 
 
     def _gen_batches(self, leak, source_id, batch_size=10000):
@@ -547,82 +597,96 @@ class DB():
 
     def _add_batches(self, batch_queue, result_queue, comms_queue, source_id):
 
-        max_attempts = 3
-
-        mongo_main_client = pymongo.MongoClient('127.0.0.1', 27000)
-        mongo_meta_client = pymongo.MongoClient('127.0.0.1', 27001)
-        mongo_main = mongo_main_client['credshed']
-        mongo_meta = mongo_meta_client['credshed']
-        unique_accounts = 0
-
         try:
-            comms_queue.put(True)
-            sleep(1)
 
-            while 1:
-                try:
-                    #errprint('[+] Getting batch')
-                    num_inserted = 0
-                    batch = batch_queue.get_nowait()
+            max_attempts = 3
 
-                    if batch is None:
-                        break
-                    else:
-                        '''
-                        timeout_value = (int(len(batch) / 500)) + (5 * max_attempts) + 1
-                        try:
-                            with concurrent.futures.ThreadPoolExecutor(max_workers=2) as thread_executor:
+            main_server = self.config['MONGO PRIMARY']['server']
+            main_port = int(self.config['MONGO PRIMARY']['port'])
+            main_db = self.config['MONGO PRIMARY']['db']
 
-                                main_thread = thread_executor.submit(self._mongo_main_add_batch, mongo_main, source_id, copy.deepcopy(batch))
-                                meta_thread = thread_executor.submit(self._mongo_meta_add_batch, mongo_meta, source_id, batch)
+            mongo_main_client = pymongo.MongoClient(main_server, main_port)
+            mongo_main = mongo_main_client[main_db]
+            unique_accounts = 0
 
-                            num_inserted = main_thread.result(timeout=timeout_value)
+            if self.use_metadata:
+                meta_server = self.config['MONGO METADATA']['server']
+                meta_port = int(self.config['MONGO METADATA']['port'])
+                meta_db = self.config['MONGO METADATA']['db']
+                mongo_meta_client = pymongo.MongoClient(meta_server, meta_port)
+                mongo_meta = mongo_meta_client[meta_db]
 
-                        except concurrent.futures.TimeoutError:
-                            #comms_queue.put('"main" mongodb thread timed out after {:,} seconds'.format(timeout_value))
-                            continue
+            try:
+                comms_queue.put(True)
+                sleep(1)
 
-                        finally:
-                            unique_accounts += num_inserted
-                        '''
+                while 1:
+                    try:
+                        #errprint('[+] Getting batch')
+                        num_inserted = 0
+                        batch = batch_queue.get_nowait()
 
-                        try:
-
-                            num_inserted = self._mongo_main_add_batch(mongo_main, source_id, copy.deepcopy(batch))
-                            self._mongo_meta_add_batch(mongo_meta, source_id, batch)
-
-                        except pymongo.errors.PyMongoError as e:
-                            error = str(e)
+                        if batch is None:
+                            break
+                        else:
+                            '''
+                            timeout_value = (int(len(batch) / 500)) + (5 * max_attempts) + 1
                             try:
-                                error += (str(e.details))
-                            except AttributeError:
-                                pass
-                            comms_queue.put('Error in _add_batches():\n{}'.format(error))
+                                with concurrent.futures.ThreadPoolExecutor(max_workers=2) as thread_executor:
 
-                        unique_accounts += num_inserted
+                                    main_thread = thread_executor.submit(self._mongo_main_add_batch, mongo_main, source_id, copy.deepcopy(batch))
+                                    meta_thread = thread_executor.submit(self._mongo_meta_add_batch, mongo_meta, source_id, batch)
+
+                                num_inserted = main_thread.result(timeout=timeout_value)
+
+                            except concurrent.futures.TimeoutError:
+                                #comms_queue.put('"main" mongodb thread timed out after {:,} seconds'.format(timeout_value))
+                                continue
+
+                            finally:
+                                unique_accounts += num_inserted
+                            '''
+
+                            try:
+
+                                num_inserted = self._mongo_main_add_batch(mongo_main, source_id, copy.deepcopy(batch))
+                                if self.use_metadata:
+                                    self._mongo_meta_add_batch(mongo_meta, source_id, batch)
+
+                            except pymongo.errors.PyMongoError as e:
+                                error = str(e)
+                                try:
+                                    error += (str(e.details))
+                                except AttributeError:
+                                    pass
+                                comms_queue.put('Error in _add_batches():\n{}'.format(error))
+
+                            unique_accounts += num_inserted
 
 
-                except queue.Empty:
-                    sleep(.1)
-                    continue
+                    except queue.Empty:
+                        sleep(.1)
+                        continue
+
+            except KeyboardInterrupt:
+                return
+
+            except Exception as e:
+                comms_queue.put('Error in _add_batches():\n{}'.format(str(e)))
+                return
+
+            finally:
+                try:
+                    #mongo_main_client.close()
+                    #mongo_meta_client.close()
+                    result_queue.put(unique_accounts)
+                    # send signal that thread is finished
+                    result_queue.put(None)
+                except:
+                    pass
 
         except KeyboardInterrupt:
-            #comms_queue.put('_add_batches() interrupted\n')
             return
-
-        except Exception as e:
-            comms_queue.put('Error in _add_batches():\n{}'.format(str(e)))
-            return
-
-        finally:
-            try:
-                #mongo_main_client.close()
-                #mongo_meta_client.close()
-                result_queue.put(unique_accounts)
-                # send signal that thread is finished
-                result_queue.put(None)
-            except:
-                pass
